+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/impl/accum/Variance.java
@@ -224,7 +224,7 @@ public class Variance extends BaseAccumulation {
             return;
         }
         int[] retShape = ArrayUtil.removeIndex(x.shape(), dimension);
-        long nOps = x.tensorssAlongDimension(dimension);
+        int nOps = x.tensorssAlongDimension(dimension);
         z = Nd4j.create(retShape);
         for( int i = 0; i < nOps; i++ ) {
             double d = Nd4j.getExecutioner().execAndReturn(opForDimension(i, dimension)).getFinalResult().doubleValue();
